Fixed circular dependency That was fatal in Cordova
